pyproject.toml
[project]
name = "vllms"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.1",
    "img2dataset>=1.47.0",
    "peft>=0.18.0",
    "rich>=14.2.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "pandas>=2.0.0",
    "Pillow>=10.0.0",
    "requests>=2.28.0",
]

[tool.setuptools.packages.find]
include = ["vlm_train*"]